shareIP
This Java saves your public IP to a text file in any folder every hour. I personally set it so that it saves to a dropbox folder to which I have a link to. I'm sure this could be edited to make use of an ftp server or something like that.